requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
google-genai==0.3.0
python-dotenv==1.0.1
pydantic==2.6.3